move new ui components into position
