--- app/b/page.tsx
@@ -1,10 +1,0 @@
-import { PageWrapper } from '@/app/b/_components/PageWrapper'
-import { Logo } from '@/components/ui/Logo'
-
-export default function Page() {
-  return (
-    <PageWrapper className="flex">
-      <Logo className="m-auto size-48" />
-    </PageWrapper>
-  )
-}
