--- app/error.tsx
@@ -1,23 +1,0 @@
-'use client'
-
-import { useEffect } from 'react'
-
-import { SunBarLoader } from '@/components/ui/SunBarLoader'
-
-export default function Error({
-  error,
-  reset,
-}: {
-  error: Error & { digest?: string }
-  reset: () => void
-}) {
-  useEffect(() => {
-    console.error(error)
-  }, [error])
-
-  return (
-    <div onClick={() => reset()}>
-      <SunBarLoader alert />
-    </div>
-  )
-}
