+++ app/layout.tsx
@@ -8,6 +8,7 @@ import { SpeedInsights } from '@vercel/speed-insights/next'
 import { Chakra_Petch, IBM_Plex_Mono, Inter } from 'next/font/google'
 import { Toaster } from 'sonner'
 
+import { AppLayout } from '@/components/AppLayout'
 import { ClientProviders } from '@/components/util/ClientProviders'
 import { cn, environment } from '@/lib/utils'
 
@@ -27,7 +28,7 @@ export const metadata: Metadata = {
 }
 
 export const viewport: Viewport = {
-  themeColor: '#17120E',
+  themeColor: '#111113',
   colorScheme: 'dark',
 }
 
@@ -72,7 +73,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
         <ClerkProvider appearance={{ baseTheme: dark }}>
           <ClientProviders>
             <Theme accentColor="orange" grayColor="slate" appearance="dark" panelBackground="solid">
-              {children}
+              <AppLayout>{children}</AppLayout>
               <Toaster position="top-right" theme="light" closeButton richColors />
             </Theme>
           </ClientProviders>
