+++ app/loading.tsx
@@ -1,5 +1,9 @@
-import { SunBarLoader } from '@/components/ui/SunBarLoader'
+import { Logo } from '@/components/ui/Logo'
 
 export default function Loading() {
-  return <SunBarLoader />
+  return (
+    <div className="flex h-full w-full bg-gray-1">
+      <Logo className="m-auto size-48 animate-pulse brightness-[.25] saturate-0" />
+    </div>
+  )
 }
