+++ app/page.tsx
@@ -1,12 +1,3 @@
-'use client'
-
-import { useEffect } from 'react'
-import { useRouter } from 'next/navigation'
-
 export default function Page() {
-  const router = useRouter()
-  useEffect(() => {
-    router.push('/c')
-  }, [router])
   return <div></div>
 }
