+++ components/message/Message.tsx
@@ -4,12 +4,12 @@ import { DropdownMenu, IconButton } from '@radix-ui/themes'
 import dynamic from 'next/dynamic'
 import { useRouter } from 'next/navigation'
 
-import { MessageEditor } from '@/app/b/c/[...slug]/_components/MessageEditor'
 import { ImageCard } from '@/components/images/ImageCard'
 import { ImageGeneratingEffect } from '@/components/images/ImageGeneratingEffect'
 import { useLightbox } from '@/components/lightbox/hooks'
 import { Marble, useMarbleProperties } from '@/components/marble-avatar/Marble'
 import { Markdown } from '@/components/message/Markdown'
+import { MessageEditor } from '@/components/message/MessageEditor'
 import { LoadingSpinner } from '@/components/ui/LoadingSpinner'
 import { Pre } from '@/components/util/Pre'
 import { getActiveJobs } from '@/convex/shared/utils'
