+++ components/pages/ChatPage.tsx
@@ -7,20 +7,20 @@ import { Authenticated, AuthLoading, Unauthenticated } from 'convex/react'
 import { useRouter } from 'next/navigation'
 import { useInView } from 'react-intersection-observer'
 
-import { PageWrapper } from '@/app/b/_components/PageWrapper'
-import { useAppendMessage } from '@/app/b/api'
-import { FilterControl } from '@/app/b/c/[...slug]/_components/FilterControl'
-import { Message } from '@/app/b/c/[...slug]/_components/Message'
-import { ChatProvider, useChat } from '@/app/b/c/[...slug]/_provider/ChatProvider'
-import { appConfig } from '@/app/b/config'
 import { Composer } from '@/components/composer/Composer'
 import { UserButtons } from '@/components/layout/UserButtons'
+import { Message } from '@/components/message/Message'
+import { FilterControl } from '@/components/pages/FilterControl'
+import { PageWrapper } from '@/components/pages/PageWrapper'
+import { ChatProvider, useChat } from '@/components/providers/ChatProvider'
 import { useShellActions } from '@/components/shell/hooks'
 import { ShellC } from '@/components/shell/Shell'
 import { Link } from '@/components/ui/Link'
 import { LinkButton2 } from '@/components/ui/LinkButton'
 import { AdminOnlyUi } from '@/components/util/AdminOnlyUi'
 import { Pre } from '@/components/util/Pre'
+import { appConfig } from '@/config/config'
+import { useAppendMessage } from '@/lib/api'
 import { useViewerDetails } from '@/lib/queries'
 import { cn } from '@/lib/utils'
