+++ components/pages/FilterControl.tsx
@@ -2,7 +2,7 @@ import * as Icons from '@phosphor-icons/react/dist/ssr'
 import { Label } from '@radix-ui/react-label'
 import { Button, Checkbox, Popover, Radio, Separator } from '@radix-ui/themes'
 
-import { useChat } from '@/app/b/c/[...slug]/_provider/ChatProvider'
+import { useChat } from '@/components/providers/ChatProvider'
 
 export const FilterControl = ({
   buttonProps,
